requirements.txt
Django==2.2.5
django-bootstrap-modal-forms==2.0.0
django-bootstrap4==1.1.1
django-filter==2.4.0
django-widget-tweaks==1.4.5
